Fixed "./gradlew publishToMavenLocal" compiler errors

+++ resources/src/iosMain/kotlin/uikit/BackgroundStyle.kt
@@ -23,6 +23,7 @@ import kotlinx.cinterop.CValue
 import kotlinx.cinterop.useContents
 import platform.CoreFoundation.CFRetain
 import platform.CoreGraphics.CGColorRef
+import platform.CoreGraphics.CGFloat
 import platform.CoreGraphics.CGPathRef
 import platform.CoreGraphics.CGPointMake
 import platform.CoreGraphics.CGRect
@@ -63,7 +64,7 @@ private fun pathForShape(shape: BackgroundStyle.Shape, bounds: CValue<CGRect>): 
             BackgroundStyle.Shape.Rectangle.Corner.BOTTOM_RIGHT -> UIRectCornerBottomRight
         }
     },
-        CGSizeMake(shape.cornerRadiusX.toDouble(), shape.cornerRadiusY.toDouble())
+        CGSizeMake(shape.cornerRadiusX.toDouble() as CGFloat, shape.cornerRadiusY.toDouble() as CGFloat)
     )
     is BackgroundStyle.Shape.Oval -> UIBezierPath.bezierPathWithOvalInRect(bounds)
 }.CGPath
@@ -97,30 +98,30 @@ private fun CAGradientLayer.applyGradientStyle(gradientStyle: GradientStyle, bou
     is GradientStyle.Linear -> {
         type = kCAGradientLayerAxial
         val startAndEndPoint = when (gradientStyle.orientation) {
-            GradientStyle.Linear.Orientation.TOP_LEFT_BOTTOM_RIGHT -> Pair(CGPointMake(0.0, 0.0), CGPointMake(1.0, 1.0))
-            GradientStyle.Linear.Orientation.TOP_RIGHT_BOTTOM_LEFT -> Pair(CGPointMake(1.0, 0.0), CGPointMake(0.0, 0.0))
-            GradientStyle.Linear.Orientation.TOP_BOTTOM -> Pair(CGPointMake(0.5, 0.0), CGPointMake(0.5, 1.0))
-            GradientStyle.Linear.Orientation.LEFT_RIGHT -> Pair(CGPointMake(0.0, 0.5), CGPointMake(1.0, 0.5))
-            GradientStyle.Linear.Orientation.BOTTOM_LEFT_TOP_RIGHT -> Pair(CGPointMake(0.0, 1.0), CGPointMake(1.0, 0.0))
-            GradientStyle.Linear.Orientation.BOTTOM_RIGHT_TOP_LEFT -> Pair(CGPointMake(1.0, 1.0), CGPointMake(0.0, 0.0))
-            GradientStyle.Linear.Orientation.BOTTOM_TOP -> Pair(CGPointMake(0.5, 1.0), CGPointMake(0.5, 0.0))
-            GradientStyle.Linear.Orientation.RIGHT_LEFT -> Pair(CGPointMake(1.0, 0.5), CGPointMake(0.0, 0.5))
+            GradientStyle.Linear.Orientation.TOP_LEFT_BOTTOM_RIGHT -> Pair(CGPointMake(0.0 as CGFloat, 0.0 as CGFloat), CGPointMake(1.0 as CGFloat, 1.0 as CGFloat))
+            GradientStyle.Linear.Orientation.TOP_RIGHT_BOTTOM_LEFT -> Pair(CGPointMake(1.0 as CGFloat, 0.0 as CGFloat), CGPointMake(0.0 as CGFloat, 0.0 as CGFloat))
+            GradientStyle.Linear.Orientation.TOP_BOTTOM -> Pair(CGPointMake(0.5 as CGFloat, 0.0 as CGFloat), CGPointMake(0.5 as CGFloat, 1.0 as CGFloat))
+            GradientStyle.Linear.Orientation.LEFT_RIGHT -> Pair(CGPointMake(0.0 as CGFloat, 0.5 as CGFloat), CGPointMake(1.0 as CGFloat, 0.5 as CGFloat))
+            GradientStyle.Linear.Orientation.BOTTOM_LEFT_TOP_RIGHT -> Pair(CGPointMake(0.0 as CGFloat, 1.0 as CGFloat), CGPointMake(1.0 as CGFloat, 0.0 as CGFloat))
+            GradientStyle.Linear.Orientation.BOTTOM_RIGHT_TOP_LEFT -> Pair(CGPointMake(1.0 as CGFloat, 1.0 as CGFloat), CGPointMake(0.0 as CGFloat, 0.0 as CGFloat))
+            GradientStyle.Linear.Orientation.BOTTOM_TOP -> Pair(CGPointMake(0.5 as CGFloat, 1.0 as CGFloat), CGPointMake(0.5 as CGFloat, 0.0 as CGFloat))
+            GradientStyle.Linear.Orientation.RIGHT_LEFT -> Pair(CGPointMake(1.0 as CGFloat, 0.5 as CGFloat), CGPointMake(0.0 as CGFloat, 0.5 as CGFloat))
         }
         startPoint = startAndEndPoint.first
         endPoint = startAndEndPoint.second
     }
     is GradientStyle.Radial -> {
         type = kCAGradientLayerRadial
-        startPoint = CGPointMake(gradientStyle.centerPoint.x.toDouble(), gradientStyle.centerPoint.y.toDouble())
+        startPoint = CGPointMake(gradientStyle.centerPoint.x.toDouble() as CGFloat, gradientStyle.centerPoint.y.toDouble() as CGFloat)
         endPoint = CGPointMake(
-            gradientStyle.centerPoint.x.toDouble() + (gradientStyle.radius / bounds.useContents { size.width }),
-            gradientStyle.centerPoint.y.toDouble() + (gradientStyle.radius / bounds.useContents { size.height })
+            gradientStyle.centerPoint.x.toDouble() as CGFloat + (gradientStyle.radius / bounds.useContents { size.width }),
+            gradientStyle.centerPoint.y.toDouble() as CGFloat + (gradientStyle.radius / bounds.useContents { size.height })
         )
     }
     is GradientStyle.Angular -> {
         type = kCAGradientLayerConic
-        startPoint = CGPointMake(gradientStyle.centerPoint.x.toDouble(), gradientStyle.centerPoint.y.toDouble())
-        endPoint = CGPointMake(1.0, gradientStyle.centerPoint.y.toDouble())
+        startPoint = CGPointMake(gradientStyle.centerPoint.x.toDouble() as CGFloat, gradientStyle.centerPoint.y.toDouble() as CGFloat)
+        endPoint = CGPointMake(1.0 as CGFloat, gradientStyle.centerPoint.y.toDouble() as CGFloat)
     }
 }
 
@@ -130,11 +131,11 @@ private fun CALayer.applyStroke(strokeStyle: BackgroundStyle.StrokeStyle, path: 
         this.path = path
         when (strokeStyle) {
             is BackgroundStyle.StrokeStyle.Stroke -> {
-                lineWidth = strokeStyle.width.toDouble()
+                lineWidth = strokeStyle.width.toDouble() as CGFloat
                 strokeColor = strokeStyle.color.uiColor.CGColor
             }
             is BackgroundStyle.StrokeStyle.None -> {
-                lineWidth = 0.0
+                lineWidth = 0.0 as CGFloat
                 strokeColor = UIColor.clearColor.CGColor
             }
         }
